+++ GenerateBoundaryPairs/GeneratePairsOfTestData.cpp
@@ -20,7 +20,7 @@ std::pair<S6, S6> GeneratePairsOfTestData::ScalePairToSpecifedSeparation(const S
    return std::make_pair(factor*s1 / deltanorm, factor*s2 / deltanorm);
 }
 
-S6 GeneratePairsOfTestData::DoReflection(const S6& s, const unsigned long n) {
+S6 GeneratePairsOfTestData::DoReflection(const S6& s, const size_t n) {
    static const std::vector< S6(*)(const S6&)> refl = S6::SetRelectionFunctions();
    return refl[n](s);
 }
@@ -34,7 +34,7 @@ std::pair<S6, S6> GeneratePairsOfTestData::GeneratePairsOfTestData::BinarySearch
    // norm doesn't work here because s2 is invalid !!!!!!!!!!
    // operator- doesn't work here because s2 is invalid !!!!!!!!!!
    S6 s6delta;
-   for (unsigned long i = 0; i < 6; ++i) s6delta[i] = s2[i] - s1[i];
+   for (size_t i = 0; i < 6; ++i) s6delta[i] = s2[i] - s1[i];
    S6 midpoint = s1 + 0.5 * s6delta;
    const bool bmid = LRL_Cell(midpoint).GetValid();
    midpoint.SetValid(bmid);
@@ -53,7 +53,7 @@ S6 GeneratePairsOfTestData::DoRandomReflection(const S6& s) {
    return DoReflection(s, rand() & 23);
 }
 
-std::pair<S6, S6> GeneratePairsOfTestData::GeneratePair_Reduced_Other_ReducesFarAway(const unsigned long numberOfZeros) {
+std::pair<S6, S6> GeneratePairsOfTestData::GeneratePair_Reduced_Other_ReducesFarAway(const size_t numberOfZeros) {
    static GenerateRandomLattice<S6> generator(seed);
    const S6 v1 = generator.randSellingReduced();
    const S6 v2 = generator.GenerateExtreme();
@@ -71,7 +71,7 @@ std::string GeneratePairsOfTestData::FormatPairOfTestS6Vectors(const S6& s1, con
    return LRL_ToString("S ", s1, "    ", id) + " " + LRL_ToString("\nS ", s2, "    ", id, "\n");
 }
 
-void GeneratePairsOfTestData::MultiGeneratePair_Reduced_Other_ReducesFarAway(const unsigned long targetReducedUnreducedPairCount) {
+void GeneratePairsOfTestData::MultiGeneratePair_Reduced_Other_ReducesFarAway(const size_t targetReducedUnreducedPairCount) {
    /*
    One point is generated. It is slightly modified, and it is reduced. If the reduced one is
    somewhat or more farther away than the modification, then the pair is kept and output.
@@ -80,7 +80,7 @@ void GeneratePairsOfTestData::MultiGeneratePair_Reduced_Other_ReducesFarAway(con
    store.SetTitle("Generate Pairs with one reduced and the other nearby but reduces far.");
    std::pair<S6, S6> output;
 
-   for (unsigned long i = 0; i < targetReducedUnreducedPairCount; ++i) {
+   for (size_t i = 0; i < targetReducedUnreducedPairCount; ++i) {
       output = GeneratePair_Reduced_Other_ReducesFarAway(0);
       const int rank = Rank(output.first);
       store.Store(rank, LRL_ToString("S", output.first, "\nS    ", output.second, "\n"));
@@ -96,25 +96,25 @@ int GeneratePairsOfTestData::Rank(const S6& s) {
    const double length = s.norm();
    double minS6 = DBL_MAX;;
 
-   unsigned long nzero = 0;
-   for (unsigned long i = 0; i < 6; ++i) {
+   size_t nzero = 0;
+   for (size_t i = 0; i < 6; ++i) {
       minS6 = std::abs(std::min(minS6, s[i]));
    }
    return (int)(log10(minS6));
 }
 
-void GeneratePairsOfTestData::GenerateRandomLatticesAndCatalogByReductionStepCount(const unsigned long n) {
+void GeneratePairsOfTestData::GenerateRandomLatticesAndCatalogByReductionStepCount(const size_t n) {
    GenerateRandomLattice<S6> grl(19191);
-   StoreResults<unsigned long, std::string> storeExtreme(3);
+   StoreResults<size_t, std::string> storeExtreme(3);
    storeExtreme.SetTitle("Extreme Unreduced S6 Generation -- >1000 cycles means reduction failed");
    storeExtreme.SetKeyLabel("cycles");
-   StoreResults<unsigned long, std::string> storesimple(3);
+   StoreResults<size_t, std::string> storesimple(3);
    storesimple.SetTitle("S6 Selling Unreduced Generation -- >1000 cycles means reduction failed");
    storesimple.SetKeyLabel("cycles");
    S6 s6in, out;
    bool b;
 
-   for (unsigned long i = 0; i < n; ++i) {
+   for (size_t i = 0; i < n; ++i) {
       s6in = grl.GenerateExtreme();
       if (s6in.GetValid()) {
          b = Selling::Reduce(s6in, out);
@@ -139,7 +139,7 @@ void GeneratePairsOfTestData::GenerateRandomLatticesAndCatalogByReductionStepCou
 double GeneratePairsOfTestData::PositiveDistance(const S6& s6) {
    double sum = 0;
    double sqsum = 0;
-   for (unsigned long i = 0; i < 6; ++i) {
+   for (size_t i = 0; i < 6; ++i) {
       sum += (s6[i] > 0.0) ? s6[i] * s6[i] : 0.0;
       sqsum += s6[i] * s6[i];
    }
@@ -147,14 +147,14 @@ double GeneratePairsOfTestData::PositiveDistance(const S6& s6) {
 }
 
 static const std::string types[] = { std::string("C1 (cI)"), std::string("C3 (cF)"), std::string("C5 (cP)"), std::string("T1 (tI)"), std::string("t2 (tI)"), std::string("t5 (tP)"), std::string("R1 (rP)"), std::string("R3 (rP)"), std::string("O1A (oF)"), std::string("O1B (oI)"), std::string("O2 (oI)"), std::string("O3 (oI)"), std::string("O4 (oS)"), std::string("O5 (oP)"), std::string("M1A (mC)"), std::string("M1B (mC)"), std::string("M2B (mC)"), std::string("M2A (mC)"), std::string("M3 (mC)"), std::string("M4 (mP)"), std::string("A1 (aP)"), std::string("A2 (aP)"), std::string("A3 (aP)"), std::string("H4 (hP)") };
-static const unsigned long avoidList[] = { 0,1,3,4,8,9 };
+static const size_t avoidList[] = { 0,1,3,4,8,9 };
 
-bool GeneratePairsOfTestData::IsGoodLatticeTypeForPerturbation(const unsigned long n) {
-   static const std::vector<unsigned long> vAvoid(avoidList, avoidList + sizeof(avoidList) / sizeof(avoidList[0]));
+bool GeneratePairsOfTestData::IsGoodLatticeTypeForPerturbation(const size_t n) {
+   static const std::vector<size_t> vAvoid(avoidList, avoidList + sizeof(avoidList) / sizeof(avoidList[0]));
    return std::find(vAvoid.begin(), vAvoid.end(), n) == vAvoid.end();
 }
 
-std::pair<S6, S6> GeneratePairsOfTestData::GenerateLatticeTypeExamplesNearDeloneTypes(const unsigned long n) {
+std::pair<S6, S6> GeneratePairsOfTestData::GenerateLatticeTypeExamplesNearDeloneTypes(const size_t n) {
    /*
    
    */
@@ -165,8 +165,8 @@ std::pair<S6, S6> GeneratePairsOfTestData::GenerateLatticeTypeExamplesNearDelone
    static const std::vector<MatS6> vDeloneTypes = Delone::LoadLatticeTypeProjectors();
 
    std::pair<S6, S6> p;
-   for (unsigned long ncycles = 0; ncycles < std::max(1UL, n); ++ncycles) {
-      for (unsigned long i = 0; i < vDeloneTypes.size(); ++i) {
+   for (size_t ncycles = 0; ncycles < std::max(size_t(1), n); ++ncycles) {
+      for (size_t i = 0; i < vDeloneTypes.size(); ++i) {
          if (!IsGoodLatticeTypeForPerturbation(i)) continue;
          p = OneLatticeType(vDeloneTypes[i]);
          const double pnorm = (p.first - p.second).norm();
@@ -189,8 +189,8 @@ std::pair<S6, S6> GeneratePairsOfTestData::GenerateLatticeTypeExamplesNearDelone
 
 std::pair<S6, S6> GeneratePairsOfTestData::OneLatticeType(const MatS6& m) {
    bool again = true;
-   unsigned long npass = 0;
-   unsigned long maxPass = 100;
+   size_t npass = 0;
+   size_t maxPass = 100;
    std::pair<S6, S6> p;
    std::pair<S6, S6> pmaxDiff;
    S6 red;
@@ -235,8 +235,8 @@ S6 GeneratePairsOfTestData::Perturb(const S6& s, const double fraction) {
    static const std::vector<MatS6> vDeloneTypes = Delone::LoadLatticeTypeProjectors();
    S6 temp;
 
-   unsigned long npass = 0;
-   unsigned long maxPass = 100;
+   size_t npass = 0;
+   size_t maxPass = 100;
    LRL_Cell_Degrees cell(s);
    const double snorm = s.norm();
    while (again && npass < maxPass) {
@@ -256,8 +256,8 @@ S6 GeneratePairsOfTestData::Perturb(const S6& s, const double fraction) {
 static int oneAttemptSeed = 1;
 
 S6 GeneratePairsOfTestData::GenerateStartingPoint(const MatS6& m) {
-   unsigned long npass = 0;
-   unsigned long maxPass = 100;
+   size_t npass = 0;
+   size_t maxPass = 100;
    S6 test = m * generator.randSellingReduced();
 
    while (!LRL_Cell(test).GetValid() && npass < maxPass) {
@@ -274,8 +274,8 @@ S6 GeneratePairsOfTestData::GenerateStartingPoint(const MatS6& m) {
 
 S6 AddOneToSingleScalar( const S6& s) {
    double mins = DBL_MAX;
-   unsigned long minIndex;
-   for (unsigned long i = 0; i < 6; ++i) {
+   size_t minIndex;
+   for (size_t i = 0; i < 6; ++i) {
       if (s[i] < mins) {
          mins = s[i];
          minIndex = i;
@@ -286,12 +286,12 @@ S6 AddOneToSingleScalar( const S6& s) {
    return s6;
 }
 
-void GeneratePairsOfTestData::StoreOnePointIfValid(StoreResults<unsigned long, std::string >& store, const S6& s6in) {
+void GeneratePairsOfTestData::StoreOnePointIfValid(StoreResults<size_t, std::string >& store, const S6& s6in) {
    if (s6in.IsValid()) {
       S6 out;
       const bool b = Selling::Reduce(s6in, out);
       if (b && out.IsValid()) {
-         const unsigned long ncycles = Selling::GetCycles();
+         const size_t ncycles = Selling::GetCycles();
          const std::string label = LRL_ToString(ncycles) + " Ex";
          store.Store(Selling::GetCycles(), GeneratePairsOfTestData::FormatPairOfTestS6Vectors(AddOneToSingleScalar(s6in), out, label));
       }
